+++ pkgs/objective_c/test/converter_test.dart
@@ -24,6 +24,10 @@ void main() {
       final obj = NSObject();
       expect(toObjCObject(obj), obj);
 
+      // NSNull.null$() is a singleton.
+      expect(toObjCObject(null), NSNull.null$());
+      expect(toNullableDartObject(NSNull.null$()), null);
+
       expect(toObjCObject(123), isA<NSNumber>());
       expect((toObjCObject(123) as NSNumber).longLongValue, 123);
       expect(toDartObject(toObjCObject(123)), isA<int>());